added option "--backup" to use parameter "--mirror" when cloning

+++ cmd/clone.go
@@ -26,6 +26,7 @@ var (
 	color             string
 	baseURL           string
 	skipArchived      bool
+	backup            bool
 	args              []string
 )
 
@@ -43,6 +44,8 @@ func init() {
 	cloneCmd.Flags().StringVarP(&namespace, "namespace", "n", "", "GHORG_GITLAB_DEFAULT_NAMESPACE - gitlab only: limits clone targets to a specific namespace e.g. --namespace=gitlab-org/security-products")
 
 	cloneCmd.Flags().BoolVar(&skipArchived, "skip-archived", false, "GHORG_SKIP_ARCHIVED skips archived repos, github/gitlab only")
+
+	cloneCmd.Flags().BoolVar(&backup, "backup", false, "GHORG_BACKUP backup mode, clone as mirror, no working copy (ignores branch parameter)")
 
 	cloneCmd.Flags().StringVarP(&baseURL, "base-url", "", "", "GHORG_SCM_BASE_URL change SCM base url, for on self hosted instances (currently gitlab only, use format of https://git.mydomain.com/api/v3)")
 }
@@ -107,6 +110,10 @@ var cloneCmd = &cobra.Command{
 
 		if cmd.Flags().Changed("skip-archived") {
 			os.Setenv("GHORG_SKIP_ARCHIVED", "true")
+		}
+
+		if cmd.Flags().Changed("backup") {
+			os.Setenv("GHORG_BACKUP", "true")
 		}
 
 		configs.GetOrSetToken()
@@ -313,40 +320,53 @@ func CloneAllRepos() {
 			repoDir := os.Getenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO") + args[0] + "_ghorg" + "/" + appName
 
 			if repoExistsLocally(repoDir) == true {
+				if os.Getenv("GHORG_BACKUP") == "true" {
+					cmd := exec.Command("git", "remote", "update")
+					cmd.Dir = repoDir
+					err := cmd.Run()
+					if err != nil {
+						infoc <- fmt.Errorf("Could not update remotes in Repo: %s Error: %v", branch, repoUrl, err)
+						return
+					}
+				} else {
+					cmd := exec.Command("git", "checkout", branch)
+					cmd.Dir = repoDir
+					err := cmd.Run()
+					if err != nil {
+						infoc <- fmt.Errorf("Could not checkout out %s, no changes made Repo: %s Error: %v", branch, repoUrl, err)
+						return
+					}
 
-				cmd := exec.Command("git", "checkout", branch)
-				cmd.Dir = repoDir
-				err := cmd.Run()
-				if err != nil {
-					infoc <- fmt.Errorf("Could not checkout out %s, no changes made Repo: %s Error: %v", branch, repoUrl, err)
-					return
-				}
+					cmd = exec.Command("git", "clean", "-f", "-d")
+					cmd.Dir = repoDir
+					err = cmd.Run()
+					if err != nil {
+						errc <- fmt.Errorf("Problem running git clean: %s Error: %v", repoUrl, err)
+						return
+					}
 
-				cmd = exec.Command("git", "clean", "-f", "-d")
-				cmd.Dir = repoDir
-				err = cmd.Run()
-				if err != nil {
-					errc <- fmt.Errorf("Problem running git clean: %s Error: %v", repoUrl, err)
-					return
-				}
+					cmd = exec.Command("git", "fetch", "-n", "origin", branch)
+					cmd.Dir = repoDir
+					err = cmd.Run()
+					if err != nil {
+						errc <- fmt.Errorf("Problem trying to fetch %v Repo: %s Error: %v", branch, repoUrl, err)
+						return
+					}
 
-				cmd = exec.Command("git", "fetch", "-n", "origin", branch)
-				cmd.Dir = repoDir
-				err = cmd.Run()
-				if err != nil {
-					errc <- fmt.Errorf("Problem trying to fetch %v Repo: %s Error: %v", branch, repoUrl, err)
-					return
-				}
-
-				cmd = exec.Command("git", "reset", "--hard", "origin/"+branch)
-				cmd.Dir = repoDir
-				err = cmd.Run()
-				if err != nil {
-					errc <- fmt.Errorf("Problem resetting %s Repo: %s Error: %v", branch, repoUrl, err)
-					return
+					cmd = exec.Command("git", "reset", "--hard", "origin/"+branch)
+					cmd.Dir = repoDir
+					err = cmd.Run()
+					if err != nil {
+						errc <- fmt.Errorf("Problem resetting %s Repo: %s Error: %v", branch, repoUrl, err)
+						return
+					}
 				}
 			} else {
-				cmd := exec.Command("git", "clone", repoUrl, repoDir)
+				args := []string{"clone", repoUrl, repoDir}
+				if os.Getenv("GHORG_BACKUP") == "true" {
+					args = append(args, "--mirror")
+				}
+				cmd := exec.Command("git", args...)
 				err := cmd.Run()
 				if err != nil {
 					errc <- fmt.Errorf("Problem trying to clone Repo: %s Error: %v", repoUrl, err)
@@ -398,6 +418,9 @@ func PrintConfigs() {
 	}
 	if os.Getenv("GHORG_SKIP_ARCHIVED") == "true" {
 		colorlog.PrintInfo("* Skip Archived : " + os.Getenv("GHORG_SKIP_ARCHIVED"))
+	}
+	if os.Getenv("GHORG_BACKUP") == "true" {
+		colorlog.PrintInfo("* Backup   : " + os.Getenv("GHORG_BACKUP"))
 	}
 	colorlog.PrintInfo("*************************************")
 	fmt.Println("")
